Move handler to separated class.
