+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/terminal/TerminalJso.java
@@ -33,26 +33,8 @@ class TerminalJso extends JavaScriptObject {
         this.open(element);
     }-*/;
 
-    public final native void attachCustomKeydownHandler() /*-{
-        this.attachCustomKeydownHandler(function (ev) {
-            var C = 67;
-            var V = 86;
-            if (ev.ctrlKey) {//todo handle not shift, alt, meta and so on maybe.
-                var selection = this.document.getSelection(),
-                    collapsed = selection.isCollapsed,
-                    isRange = typeof collapsed === 'boolean' ? !collapsed : selection.type === 'Range';
-
-                    //Ctrl + C
-                    if (ev.keyCode === C && isRange) {
-                        return false;
-                    }
-
-                //Ctrl + V
-                if (ev.keyCode === V) {
-                    return false;
-                }
-            }
-        });
+    public final native void attachCustomKeyDownHandler(JavaScriptObject customKeyDownHandler) /*-{
+        this.attachCustomKeydownHandler(customKeyDownHandler);
     }-*/;
 
     public final native Element getElement() /*-{
